+++ db/log_writer.h
@@ -1,4 +1,4 @@
-//  Copyright (c) 2013, Facebook, Inc.  All rights reserved.
+//  Copyright (c) 2011-present, Facebook, Inc.  All rights reserved.
 //  This source code is licensed under the BSD-style license found in the
 //  LICENSE file in the root directory of this source tree. An additional grant
 //  of patent rights can be found in the PATENTS file in the same directory.
@@ -83,7 +83,7 @@ class Writer {
 
  private:
   unique_ptr<WritableFileWriter> dest_;
-  int block_offset_;       // Current offset in block
+  size_t block_offset_;       // Current offset in block
   uint64_t log_number_;
   bool recycle_log_files_;
