+++ db/version_set_test.cc
@@ -1,4 +1,4 @@
-//  Copyright (c) 2013, Facebook, Inc.  All rights reserved.
+//  Copyright (c) 2011-present, Facebook, Inc.  All rights reserved.
 //  This source code is licensed under the BSD-style license found in the
 //  LICENSE file in the root directory of this source tree. An additional grant
 //  of patent rights can be found in the PATENTS file in the same directory.
@@ -23,7 +23,7 @@ class GenerateLevelFilesBriefTest : public testing::Test {
   GenerateLevelFilesBriefTest() { }
 
   ~GenerateLevelFilesBriefTest() {
-    for (unsigned int i = 0; i < files_.size(); i++) {
+    for (size_t i = 0; i < files_.size(); i++) {
       delete files_[i];
     }
   }
